requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
exceptiongroup==1.1.2
Flask==2.3.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.2.0
packaging==23.1
plotly==5.15.0
psycopg2==2.9.6
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tenacity==8.2.2
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==3.8.6
Werkzeug==2.3.6
wsproto==1.2.0
